feat: refactor chat component to use ConvMessageList for improved message handling and infinite scrolling
